refactor: move cull face mode control to object not layer
