+++ packages/core/src/layers/volume_layer.ts
@@ -12,7 +12,6 @@ export class VolumeLayer extends Layer {
 
     this.addObject(renderable);
     this.setState("ready");
-    this.cullFaceMode = "back";
   }
 
   public update() {
